Revert ProHandlerBuilder, per review
