--- outwatch/src/main/scala/outwatch/dom/ProHandlerBuilder.scala
@@ -1,27 +1,0 @@
-package outwatch.dom
-
-import monix.reactive.{Observer, Observable}
-import scala.concurrent.Future
-import monix.execution.Cancelable
-import monix.reactive.observers.Subscriber
-import outwatch.ReactiveConnectable
-import monix.execution.Scheduler
-import monix.execution.Ack
-
-
-private[dom] object ProHandlerBuilder {
-  def apply[I,O](observer: Observer[I], observable: Observable[O]): ProHandler[I,O] = new Observable[O] with Observer[I] {
-    override def onNext(elem: I): Future[Ack] = observer.onNext(elem)
-    override def onError(ex: Throwable): Unit = observer.onError(ex)
-    override def onComplete(): Unit = observer.onComplete()
-    override def unsafeSubscribeFn(subscriber: Subscriber[O]): Cancelable = observable.unsafeSubscribeFn(subscriber)
-  }
-
-  def connectable[I,O](observer: Observer[I] with ReactiveConnectable, observable: Observable[O]): ProHandler[I,O] with ReactiveConnectable = new Observable[O] with Observer[I] with ReactiveConnectable {
-    override def onNext(elem: I): Future[Ack] = observer.onNext(elem)
-    override def onError(ex: Throwable): Unit = observer.onError(ex)
-    override def onComplete(): Unit = observer.onComplete()
-    override def connect()(implicit scheduler: Scheduler): Cancelable = observer.connect()
-    override def unsafeSubscribeFn(subscriber: Subscriber[O]): Cancelable = observable.unsafeSubscribeFn(subscriber)
-  }
-}
