+++ outwatch/src/main/scala/outwatch/dom/RichHandler.scala
@@ -11,12 +11,12 @@ trait RichHandlerOps {
       val redirected = self
         .redirect[S](_.withLatestFrom(self.startWith(Seq(seed))){ case (a, b) => write(b, a) })
 
-      ProHandlerBuilder.connectable(redirected, self.map(read))
+      ProHandler.connectable(redirected, self.map(read))
     }
 
-    def mapHandler[T2](write: T2 => T)(read: T => T2): Handler[T2] = ProHandlerBuilder(self.contramap(write), self.map(read))
-    def collectHandler[T2](write: PartialFunction[T2, T])(read: PartialFunction[T, T2]): Handler[T2] = ProHandlerBuilder(self.redirectCollect(write), self.collect(read))
-    def filterHandler(write: T => Boolean)(read: T => Boolean): Handler[T] = ProHandlerBuilder(self.redirectFilter(write), self.filter(read))
-    def transformHandler[T2](write: Observable[T2] => Observable[T])(read: Observable[T] => Observable[T2]): Handler[T2] with ReactiveConnectable = ProHandlerBuilder.connectable(self.redirect(write), read(self))
+    def mapHandler[T2](write: T2 => T)(read: T => T2): Handler[T2] = ProHandler(self.contramap(write), self.map(read))
+    def collectHandler[T2](write: PartialFunction[T2, T])(read: PartialFunction[T, T2]): Handler[T2] = ProHandler(self.redirectCollect(write), self.collect(read))
+    def filterHandler(write: T => Boolean)(read: T => Boolean): Handler[T] = ProHandler(self.redirectFilter(write), self.filter(read))
+    def transformHandler[T2](write: Observable[T2] => Observable[T])(read: Observable[T] => Observable[T2]): Handler[T2] with ReactiveConnectable = ProHandler.connectable(self.redirect(write), read(self))
   }
 }
